requirements.txt
tensorflow==2.15.0
numpy==1.26.4
matplotlib==3.8.2
scikit-learn==1.4.0
pillow==10.2.0
seaborn==0.13.2
flask==3.0.2
opencv-python==4.9.0.80
